Use the app folder for routing
