+++ src/app/(protected)/(main)/chat/[id]/page.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Page() {
+  return <div>Dynamic Page under chat</div>;
+}
+
+export default Page;
