+++ src/app/(protected)/(main)/photos/page.module.css
