+++ src/app/(protected)/layout.jsx
@@ -1,8 +1,12 @@
-import React from "react";
-import RequireAuth from "../../components/RequireAuth";
+import { Navigate, Outlet } from "react-router-dom";
+import LoadingPage from "../../components/LoadingPage";
+import { useUser } from "../../contexts/AuthContext";
 
-function Layout() {
-  return <RequireAuth />;
+export default function Layout() {
+  const [user, loading] = useUser();
+  const isLoggedIn = user !== null;
+
+  if (loading) return <LoadingPage />;
+  if (!isLoggedIn) return <Navigate to="/login" />;
+  return <Outlet />;
 }
-
-export default Layout;
